pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "megakernels"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "transformers==4.48.3",
    "pydra-config>=0.0.13",
    "accelerate",
    "tabulate",
    "tqdm",
    "matplotlib",
    "einops",
    "pyright",
    "openai",
    "psutil",
    "art",
    "pybind11",
]

[tool.setuptools]
packages = {find = {}}